requirements.txt
pygame-ce==2.5.5
pygame_gui==0.6.14
moderngl
numpy
bezier
